persephone_api/pyproject.toml
[build-system]
requires = ["setuptools>=67", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "persephone-api"
version = "0.1.0"
description = "MVP FastAPI backend for Persephone platform"
authors = [{name = "Persephone"}]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "fastapi[standard]>=0.110",
  "uvicorn[standard]>=0.27",
  "pydantic-settings>=2.2",
  "sqlmodel>=0.0.16",
  "sqlalchemy>=2.0",
  "alembic>=1.13",
  "psycopg[binary]>=3.1",
  "redis>=5.0",
  "httpx>=0.27",
  "boto3>=1.34",
  "prometheus-fastapi-instrumentator>=6.1",
  "python-json-logger>=2.0",
  "passlib[bcrypt]>=1.7"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "pytest-asyncio>=0.23",
  "httpx>=0.27",
  "ruff>=0.3",
  "black>=24.3",
  "mypy>=1.9"
]

[tool.black]
line-length = 88
target-version = ["py312"]

[tool.ruff]
line-length = 88
select = ["E", "F", "I", "UP", "B", "C4", "Q"]
ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
pythonpath = ["persephone_api"]

[tool.mypy]
python_version = "3.12"
strict = false
warn_return_any = true
warn_unused_configs = true
